DIRECTOR: Fix D7 MCsL libResourceId truncated to uint16
The cast-library resource id (the chunk-map resource index of a cast
library's config, used to match CASt members to their owning library) is
a 32-bit field, but it was read as uint16 in two places:

  * Movie::loadCastLibMapping() read it as two uint16s (discarding the
    high word as "unknown"), and
  * RIFXArchive::readCast() took the CAS* parentIndex as a uint16 param.

In large D7 movies this index can exceed 0xFFFF. In Physikus' Optik1.dxr
the internal "Text" library has libResourceId 0x00010402 (66562); both
truncations turned it into 0x0402 (1026), colliding with the "Animation"
library's genuine 1026. As a result loadCast() loaded the union of both
libraries' members into each cast, and the overlapping ids (1..49) were
loaded twice, triggering "Multiple cast members with ID N, overwriting"
and leaving each library polluted with the other's members.

Read the field as a full uint32 in Movie::loadCastLibMapping() and widen
RIFXArchive::readCast(), Cast::_libResourceId, the Cast ctor,
Archive::getFirstResource(tag, parentId) (so the Lctx lookup is not
truncated either) and Movie::getCastByLibResourceID() to uint32.
Libraries with a high word of 0 (the common case) are unaffected.

+++ engines/director/archive.cpp
@@ -143,7 +143,7 @@ Common::SeekableReadStreamEndian *Archive::getFirstResource(uint32 tag) {
 	return getResource(tag, getResourceIDList(tag)[0]);
 }
 
-Common::SeekableReadStreamEndian *Archive::getFirstResource(uint32 tag, uint16 parentId) {
+Common::SeekableReadStreamEndian *Archive::getFirstResource(uint32 tag, uint32 parentId) {
 	return getResource(tag, getResourceIDList(tag)[0]);
 }
 
@@ -1022,7 +1022,7 @@ bool RIFXArchive::readAfterburnerMap(Common::SeekableReadStreamEndian &stream, u
 	return true;
 }
 
-void RIFXArchive::readCast(Common::SeekableReadStreamEndian &casStream, uint16 libResourceId) {
+void RIFXArchive::readCast(Common::SeekableReadStreamEndian &casStream, uint32 libResourceId) {
 	int castTag = MKTAG('C', 'A', 'S', 't');
 
 	uint casSize = casStream.size() / 4;
@@ -1100,7 +1100,7 @@ Common::SeekableReadStreamEndian *RIFXArchive::getFirstResource(uint32 tag, bool
 	return getResource(tag, getResourceIDList(tag)[0], fileEndianness);
 }
 
-Common::SeekableReadStreamEndian *RIFXArchive::getFirstResource(uint32 tag, uint16 parentId) {
+Common::SeekableReadStreamEndian *RIFXArchive::getFirstResource(uint32 tag, uint32 parentId) {
 	if (!_keyData.contains(tag))
 		return nullptr;
 	if (!_keyData[tag].contains(parentId))
